Remove BaseTestCaseWithDatastoreAccess -> BackDoor dependency
Created a new class BaseTestCaseWithBackDoorApiAccess for tests explicitly needing dependency to BackDoor.
