+++ src/test/java/teammates/test/cases/BaseTestCaseWithDatastoreAccess.java
@@ -18,10 +18,9 @@ import teammates.common.util.Const;
 import teammates.common.util.JsonUtils;
 import teammates.common.util.StringHelper;
 import teammates.common.util.ThreadHelper;
-import teammates.test.driver.BackDoor;
 
 /**
- * Base class for all test cases which are allowed to access the Datastore via {@link BackDoor}.
+ * Base class for all test cases which are allowed to access the Datastore.
  */
 public abstract class BaseTestCaseWithDatastoreAccess extends BaseTestCase {
 
@@ -171,9 +170,7 @@ public abstract class BaseTestCaseWithDatastoreAccess extends BaseTestCase {
         }
     }
 
-    protected AccountAttributes getAccount(AccountAttributes account) {
-        return BackDoor.getAccount(account.googleId);
-    }
+    protected abstract AccountAttributes getAccount(AccountAttributes account);
 
     private void equalizeIrrelevantData(AccountAttributes expected, AccountAttributes actual) {
         // Ignore time field as it is stamped at the time of creation in testing
@@ -190,51 +187,37 @@ public abstract class BaseTestCaseWithDatastoreAccess extends BaseTestCase {
         }
     }
 
-    protected CommentAttributes getComment(CommentAttributes comment) {
-        throw new UnsupportedOperationException("Method not used");
-    }
+    protected abstract CommentAttributes getComment(CommentAttributes comment);
 
-    protected CourseAttributes getCourse(CourseAttributes course) {
-        return BackDoor.getCourse(course.getId());
-    }
+    protected abstract CourseAttributes getCourse(CourseAttributes course);
 
     private void equalizeIrrelevantData(CourseAttributes expected, CourseAttributes actual) {
         // Ignore time field as it is stamped at the time of creation in testing
         expected.createdAt = actual.createdAt;
     }
 
-    protected FeedbackQuestionAttributes getFeedbackQuestion(FeedbackQuestionAttributes fq) {
-        return BackDoor.getFeedbackQuestion(fq.courseId, fq.feedbackSessionName, fq.questionNumber);
-    }
+    protected abstract FeedbackQuestionAttributes getFeedbackQuestion(FeedbackQuestionAttributes fq);
 
     private void equalizeIrrelevantData(FeedbackQuestionAttributes expected, FeedbackQuestionAttributes actual) {
         expected.setId(actual.getId());
     }
 
-    protected FeedbackResponseCommentAttributes getFeedbackResponseComment(FeedbackResponseCommentAttributes frc) {
-        throw new UnsupportedOperationException("Method not used");
-    }
+    protected abstract FeedbackResponseCommentAttributes getFeedbackResponseComment(FeedbackResponseCommentAttributes frc);
 
-    protected FeedbackResponseAttributes getFeedbackResponse(FeedbackResponseAttributes fr) {
-        return BackDoor.getFeedbackResponse(fr.feedbackQuestionId, fr.giver, fr.recipient);
-    }
+    protected abstract FeedbackResponseAttributes getFeedbackResponse(FeedbackResponseAttributes fr);
 
     private void equalizeIrrelevantData(FeedbackResponseAttributes expected, FeedbackResponseAttributes actual) {
         expected.setId(actual.getId());
     }
 
-    protected FeedbackSessionAttributes getFeedbackSession(FeedbackSessionAttributes fs) {
-        return BackDoor.getFeedbackSession(fs.getCourseId(), fs.getFeedbackSessionName());
-    }
+    protected abstract FeedbackSessionAttributes getFeedbackSession(FeedbackSessionAttributes fs);
 
     private void equalizeIrrelevantData(FeedbackSessionAttributes expected, FeedbackSessionAttributes actual) {
         expected.setRespondingInstructorList(actual.getRespondingInstructorList());
         expected.setRespondingStudentList(actual.getRespondingStudentList());
     }
 
-    protected InstructorAttributes getInstructor(InstructorAttributes instructor) {
-        return BackDoor.getInstructorByEmail(instructor.email, instructor.courseId);
-    }
+    protected abstract InstructorAttributes getInstructor(InstructorAttributes instructor);
 
     private void equalizeIrrelevantData(InstructorAttributes expected, InstructorAttributes actual) {
         // pretend keys match because the key is generated only before storing into database
@@ -243,9 +226,7 @@ public abstract class BaseTestCaseWithDatastoreAccess extends BaseTestCase {
         }
     }
 
-    protected StudentAttributes getStudent(StudentAttributes student) {
-        return BackDoor.getStudent(student.course, student.email);
-    }
+    protected abstract StudentAttributes getStudent(StudentAttributes student);
 
     private void equalizeIrrelevantData(StudentAttributes expected, StudentAttributes actual) {
         // For these fields, we consider null and "" equivalent.
@@ -279,9 +260,7 @@ public abstract class BaseTestCaseWithDatastoreAccess extends BaseTestCase {
         assertEquals(Const.StatusCodes.BACKDOOR_STATUS_SUCCESS, backDoorOperationStatus);
     }
 
-    protected String doRemoveAndRestoreDataBundle(DataBundle testData) {
-        return BackDoor.removeAndRestoreDataBundle(testData);
-    }
+    protected abstract String doRemoveAndRestoreDataBundle(DataBundle testData);
 
     protected void putDocuments(DataBundle testData) {
         int retryLimit = OPERATION_RETRY_COUNT;
@@ -295,8 +274,6 @@ public abstract class BaseTestCaseWithDatastoreAccess extends BaseTestCase {
         assertEquals(Const.StatusCodes.BACKDOOR_STATUS_SUCCESS, backDoorOperationStatus);
     }
 
-    protected String doPutDocuments(DataBundle testData) {
-        return BackDoor.putDocuments(testData);
-    }
+    protected abstract String doPutDocuments(DataBundle testData);
 
 }
